+++ example/lib/views/page_view_renderer_page.dart
@@ -7,8 +7,11 @@ import 'package:flutter/material.dart';
 //Cubit or Bloc or Controller for the view to assign
 //the QuestionnaireRendererController instance.
 class CubitOrBlocOrController {
-  QuestionnaireRendererController? rendererController;
+  RendererQuestionnaireController rendererController =
+      RendererQuestionnaireController();
 }
+
+class CustomQuestionnaireControlller extends RendererQuestionnaireController {}
 
 class PageViewExamplePage extends StatelessWidget {
   final Questionnaire questionnaire;
@@ -25,11 +28,10 @@ class PageViewExamplePage extends StatelessWidget {
           IconButton(
             onPressed: () {
               final generatedQuestionnaireResponse =
-                  controller.rendererController
-                      ?.getGeneratedQuestionnaireResponse();
+                  controller.rendererController.generateQuestionnaireResponse();
 
               //Do something with the generated response
-              log(generatedQuestionnaireResponse?.toString() ?? "No response");
+              log(generatedQuestionnaireResponse.toString());
             },
             icon: Icon(Icons.check),
           ),
@@ -37,9 +39,7 @@ class PageViewExamplePage extends StatelessWidget {
       ),
       body: QuestionnairePageViewRenderer(
         questionnaire: questionnaire,
-        getRendererControllerInstance:
-            (QuestionnaireRendererController controller) =>
-                this.controller.rendererController = controller,
+        rendererController: controller.rendererController,
       ),
     );
   }