+++ example/lib/views/slivers_view_example_page.dart
@@ -7,7 +7,8 @@ import 'package:flutter/material.dart';
 //Cubit or Bloc or Controller for the view to assign
 //the QuestionnaireRendererController instance.
 class CubitOrBlocOrController {
-  QuestionnaireRendererController? rendererController;
+  RendererQuestionnaireController rendererController =
+      RendererQuestionnaireController(sliversInitialIndex: 3);
 }
 
 class SliversViewExamplePage extends StatelessWidget {
@@ -25,11 +26,10 @@ class SliversViewExamplePage extends StatelessWidget {
           IconButton(
             onPressed: () {
               final generatedQuestionnaireResponse =
-                  controller.rendererController
-                      ?.getGeneratedQuestionnaireResponse();
+                  controller.rendererController.generateQuestionnaireResponse();
 
               //Do something with the generated response
-              log(generatedQuestionnaireResponse?.toString() ?? "No response");
+              log(generatedQuestionnaireResponse.toString());
             },
             icon: Icon(Icons.check),
           ),
@@ -37,9 +37,7 @@ class SliversViewExamplePage extends StatelessWidget {
       ),
       body: QuestionnaireSliversViewRenderer(
         questionnaire: questionnaire,
-        getRendererControllerInstance:
-            (QuestionnaireRendererController controller) =>
-                this.controller.rendererController = controller,
+        rendererController: controller.rendererController,
         groupItemBuilder: (
           index,
           isLastItem,