+++ lib/src/ui/components/slivers/sliver_base_decorator.dart
@@ -18,7 +18,7 @@ final class SliverBaseDecorator extends BaseDecorator {
       decoration: BoxDecoration(
         color: useNotImplementedStyle
             ? Theme.of(context).colorScheme.errorContainer
-            : Theme.of(context).colorScheme.surface,
+            : null,
         borderRadius: BorderRadius.vertical(
           bottom: Radius.circular(roundBottomBorder ? 7 : 0),
         ),